refactor: split frame refresh and info overlay out of file_lightbar_render.go
