+++ internal/menu/file_lightbar_render.go
@@ -2,13 +2,11 @@ package menu
 
 import (
 	"fmt"
-	"io"
 	"strings"
 	"time"
 
 	"github.com/ViSiON-3/vision-3-bbs/internal/ansi"
 	"github.com/ViSiON-3/vision-3-bbs/internal/terminalio"
-	"github.com/ViSiON-3/vision-3-bbs/internal/user"
 )
 
 // fileLightbar rendering: file rows, separator, command bar, page indicator.
@@ -253,133 +251,6 @@ func (lb *fileLightbar) renderTop() error {
 	curPage, calcTotalPages := lb.calculatePageInfo()
 	processed := ansi.ReplacePipeCodes(processFileListPlaceholders(lb.topTemplateBytes, curPage, calcTotalPages, len(lb.allFiles), lb.fconfpath))
 	return terminalio.WriteProcessedBytes(lb.terminal, processed, lb.outputMode)
-}
-
-// lbFrame tracks the fileLightbar's smart-refresh state across run() loop
-// iterations — what was last rendered — so refreshFrame can redraw only the
-// regions that changed since then.
-type lbFrame struct {
-	prevSelectedIndex int
-	prevTopIndex      int
-	prevCmdIndex      int
-	prevPage          int
-	needFullRedraw    bool
-}
-
-// refreshFrame redraws whichever screen regions changed since the last call:
-// everything, if f.needFullRedraw is set; the whole footer and file area, if
-// the viewport scrolled; or just the old/new selected rows (plus the top
-// line, if the page changed) when only the selection moved within the same
-// viewport. The command bar is redrawn separately whenever cmdIndex changed.
-// calculatePageInfo is computed once per call and reused for both the
-// page-changed check and the trailing prevPage update — it previously ran
-// twice per keystroke.
-func (lb *fileLightbar) refreshFrame(f *lbFrame) error {
-	curPage, _ := lb.calculatePageInfo()
-
-	if f.needFullRedraw {
-		if err := lb.renderFull(); err != nil {
-			return err
-		}
-		f.needFullRedraw = false
-	} else if lb.topIndex != f.prevTopIndex {
-		// Viewport scrolled — full redraw of all regions to prevent overlap.
-		if err := lb.renderTop(); err != nil {
-			return err
-		}
-		if err := lb.renderFileArea(); err != nil {
-			return err
-		}
-		if err := lb.renderSeparator(); err != nil {
-			return err
-		}
-		if err := lb.renderCmdBar(); err != nil {
-			return err
-		}
-		if err := lb.renderPageIndicator(); err != nil {
-			return err
-		}
-	} else if lb.selectedIndex != f.prevSelectedIndex {
-		// Same viewport, selection changed — redraw old/new rows; redraw TOP if page changed.
-		if curPage != f.prevPage {
-			if err := lb.renderTop(); err != nil {
-				return err
-			}
-		}
-		if f.prevSelectedIndex >= 0 && f.prevSelectedIndex < len(lb.allFiles) {
-			if row, h := lb.screenRowForFile(f.prevSelectedIndex); row >= 0 {
-				if err := lb.writeFileRow(row, f.prevSelectedIndex, false, h); err != nil {
-					return err
-				}
-			}
-		}
-		if row, h := lb.screenRowForFile(lb.selectedIndex); row >= 0 {
-			if err := lb.writeFileRow(row, lb.selectedIndex, true, h); err != nil {
-				return err
-			}
-		}
-	}
-	if lb.cmdIndex != f.prevCmdIndex {
-		if err := lb.renderCmdBar(); err != nil {
-			return err
-		}
-	}
-
-	f.prevSelectedIndex = lb.selectedIndex
-	f.prevTopIndex = lb.topIndex
-	f.prevCmdIndex = lb.cmdIndex
-	f.prevPage = curPage
-	return nil
-}
-
-// showFileInfo displays the "i" info overlay for the selected file and
-// blocks for a keypress to dismiss it. It follows the same signal
-// convention as handleNavKey/action handlers: refresh tells run() to set
-// needFullRedraw; exit tells run() to return (result, action, err)
-// immediately (used for the LOGOFF/error paths out of the blocking
-// ReadKey call).
-func (lb *fileLightbar) showFileInfo() (refresh bool, exit bool, result *user.User, action string, err error) {
-	if len(lb.allFiles) == 0 {
-		return false, false, nil, "", nil
-	}
-	sel := lb.allFiles[lb.selectedIndex]
-	descLines := formatDIZLines(sel.Description, dizMaxWidth, dizMaxLines)
-
-	_ = terminalio.WriteProcessedBytes(lb.terminal, []byte(ansi.ClearScreen()), lb.outputMode)
-
-	d1 := fmt.Sprintf("|15Filename  : |07%s\r\n", sel.Filename)
-	d2 := fmt.Sprintf("|15Size      : |07%s\r\n", compactFileSize(sel.Size))
-	_ = lb.writePipe(d1)
-	_ = lb.writePipe(d2)
-
-	for i, dl := range descLines {
-		var dLine string
-		if i == 0 {
-			dLine = fmt.Sprintf("|15Desc      : |07%s\r\n", dl)
-		} else {
-			dLine = fmt.Sprintf("|07            %s\r\n", dl)
-		}
-		_ = lb.writePipe(dLine)
-	}
-	if len(descLines) == 0 {
-		_ = lb.writePipe("|15Desc      : |07(none)\r\n")
-	}
-
-	d3 := fmt.Sprintf("|15Uploaded  : |07%s\r\n", sel.UploadedAt.Format("01/02/2006 15:04"))
-	d4 := fmt.Sprintf("|15Uploader  : |07%s\r\n", sel.UploadedBy)
-	d5 := fmt.Sprintf("|15Downloads : |07%d\r\n", sel.DownloadCount)
-	_ = lb.writePipe(d3)
-	_ = lb.writePipe(d4)
-	_ = lb.writePipe(d5)
-
-	_ = lb.writePipe("\r\n|08Press any key to return...|07")
-	if _, waitErr := lb.ih.ReadKey(); waitErr != nil {
-		if logoffIfDisconnected(waitErr) {
-			return false, true, nil, "LOGOFF", io.EOF
-		}
-		return false, true, nil, "", waitErr
-	}
-	return true, false, nil, "", nil
 }
 
 func (lb *fileLightbar) renderFull() error {
